Extract migrate functions
